Improve scripts to account for max witness push sizes
The input scriptCode needs to be provided on the stack. Since it can be
as large as 10,000 bytes, and witness pushes can be at most 520 bytes
each, we have to provide enough space for users to push the scriptCode
on the stack. Since we have to do this anyway, and usually the
scriptCode is a lot smaller than 10kB, we can now use this space to try
and spread out the scriptCode into pushes of 80 bytes so that we don't
exceed the standardness rule of 80-byte pushes max.

Additionally, some pushes of our burn covenant also exceeded 80 bytes
(not 520), so we provided space for them to be spread out over multiple
pushes of 80 bytes max each so that in the regular case, our
transactioins can comply with standardness rules.